# docs/swagger.yaml
basePath: /v1
definitions:
  model.APIError:
    properties:
      error_message:
        type: string
      error_status:
        type: integer
    type: object
  model.DeviceClass:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      vendor_class:
        type: string
    type: object
  model.DeviceClassForm:
    properties:
      name:
        type: string
      vendor_class:
        type: string
    type: object
  model.Group:
    properties:
      bootdisk:
        type: string
      bootmethod:
        type: string
      callbackurl:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      device:
        type: string
      dns:
        type: string
      gateway:
        type: string
      host:
        items:
          $ref: '#/definitions/model.Host'
        type: array
      id:
        type: integer
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      netmask:
        type: string
      ntp:
        type: string
      options:
        additionalProperties:
          type: string
        type: object
      syslog:
        type: string
      updated_at:
        type: string
      vlan:
        type: string
    type: object
  model.GroupRequest:
    properties:
      bootdisk:
        type: string
      bootmethod:
        type: string
      callbackurl:
        type: string
      device:
        type: string
      dns:
        type: string
      gateway:
        type: string
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      netmask:
        type: string
      ntp:
        type: string
      options:
        additionalProperties:
          type: string
        type: object
      password:
        description: |-
          Password is the ESXi root password, in the clear on the way in. It is
          encrypted before it reaches the database and never sent back out.
        type: string
      syslog:
        type: string
      vlan:
        type: string
    type: object
  model.Host:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      domain:
        type: string
      expires_at:
        type: string
      first_seen:
        type: string
      group:
        $ref: '#/definitions/model.Group'
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: integer
      ip:
        type: string
      ks:
        type: string
      last_seen:
        type: string
      last_seen_relay:
        description: DHCP parameters
        type: string
      mac:
        type: string
      missing_options:
        type: string
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: VMware Imaging Appliances written in GO with full HTTP-REST API
  title: go-via
  version: "0.1"
paths:
  /device_classes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DeviceClass'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get all device classes
      tags:
      - device_classes
    post:
      consumes:
      - application/json
      parameters:
      - description: Add an device class
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.DeviceClassForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Create a new device class
      tags:
      - device_classes
  /device_classes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Remove an existing device class
      tags:
      - device_classes
    get:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get an existing device class
      tags:
      - device_classes
    patch:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an ip device class
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.DeviceClassForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Update an existing device class
      tags:
      - device_classes
  /device_classes/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fields to search for
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.DeviceClass'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Search for an device class
      tags:
      - device_classes
  /groups:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Group'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Add ip group
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.GroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Remove an existing group
      tags:
      - groups
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get an existing group
      tags:
      - groups
    patch:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an group
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.GroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Update an existing group
      tags:
      - groups
...
swagger: "2.0"